refactored ExpandedView and only render buttons when page is not zoomed
